Refactor: 잡담방 관련 구조개편, 이름 변경 chat -> talk #116
